update db migration statements to meet the requirement

+++ migrations/sqls/20220108191038-orders-table-up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE orders (
     id SERIAL PRIMARY KEY,
     status VARCHAR(15) NOT NULL,
-    user_id bigint NOT NULL REFERENCES users(id)
+    user_id integer NOT NULL REFERENCES users(id)
 );
