+++ migrations/sqls/20220108191140-order-products-table-up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE order_products (
     id SERIAL PRIMARY KEY,
     quantity integer NOT NULL CONSTRAINT positive_quantity CHECK (quantity > 0),
-    order_id bigint NOT NULL REFERENCES orders(id),
-    product_id bigint NOT NULL REFERENCES products(id)
+    order_id integer NOT NULL REFERENCES orders(id),
+    product_id integer NOT NULL REFERENCES products(id)
 );
